Bundle web server for local previewing
This enables users without Apache installed to preview a build locally. It
uses Sinatra, so requires that installed. Activate it with `rake browse` to
view the current build (in `out/`).

Sinatra can't handle content negotiation natively, so HTML and CSS files with
the same prefix clashed. As a workaround, CSS files now live in `out/c/`.