XSL: Stop using lookup table to translate 'ref.*' IDs.
Links to top-level pages in the reference section will have an ID of the form 'ref.class', where 'class' is the name of the class/module being summarised. Each paragraph on such pages will have an ID of the form 'ref.class-method', where 'method' is the method documented. For example, the URL for the Bignum page is '/ref/bignum', and the URL for Class.inherited is '/ref/class#inherited'.